Fix account activation bug.
Account activation link can now be clicked multiple times and will
always activate the account and redirect to a login page with a
success message (django.contrib.messages).

We don't want to log users in automatically on activation if the
activation email works multiple times.  This would make the activation
link a permanent substitute for a password, which weakens security.

Closes #21.